src/test/java/com/vapasi/springdemo/controller/MovieControllerTest.java
package com.vapasi.springdemo.controller;


import com.vapasi.springdemo.service.MovieService;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;


@WebMvcTest(controllers = MovieController.class)
public class MovieControllerTest {
    @Autowired
    private MockMvc mockMvc;
    @MockBean
    private MovieService movieService;

    @Test
    public void shouldReturnOkForGetMovies() throws Exception{
        mockMvc.perform(get("/api/v1/movies/")
                        .accept(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk());
        verify(movieService, times(1)).getMovies();
    }
}
